requirements.txt
streamlit
numpy
matplotlib
pyyaml
h5py
